api/requirements.txt
# Core
Django==3.1.12
djangorestframework==3.11.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.6

# Testing
pytest==6.2.2
pytest-cov==4.1.0

# Other utility libraries
python-dotenv==1.0.1
requests==2.26.0
pytz==2023.3.post1
sqlparse==0.2.4
Werkzeug==2.3.0
colorlog==6.9.0
tqdm==4.66.1
scipy==1.13.1
graphene-django==3.2.2
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
promise==2.3
text-unidecode==1.3
